CV-NLP/RealTimeDepthEstimation/requirements.txt
Flask==2.3.0
opencv-python==4.10.0
numpy==1.23.5
Pillow==9.4.0
torch==2.0.0
torchvision==0.15.0
